refactor package structure to meet original intention
