+++ app/src/main/java/com/survivalcoding/network_apps/conference/data/datasource/ConferenceURL.kt
@@ -1,0 +1,8 @@
+package com.survivalcoding.network_apps.conference.data.datasource
+
+class ConferenceURL {
+    companion object {
+        const val BASE_URL = "https://raw.githubusercontent.com/junsuk5/mock_json/main/"
+        const val CONFERENCE_URL = "conferences.json"
+    }
+}
