--- app/src/main/java/com/survivalcoding/network_apps/feature_basic/data/conference/datasource/ConferenceURL.kt
@@ -1,8 +1,0 @@
-package com.survivalcoding.network_apps.feature_basic.data.conference.datasource
-
-class ConferenceURL {
-    companion object {
-        const val BASE_URL = "https://raw.githubusercontent.com/junsuk5/mock_json/main/"
-        const val CONFERENCE_URL = "conferences.json"
-    }
-}
