+++ app/src/main/java/com/survivalcoding/network_apps/feature_basic/data/datasource/TodoDataSource.kt
@@ -1,0 +1,7 @@
+package com.survivalcoding.network_apps.feature_basic.data.datasource
+
+import com.survivalcoding.network_apps.feature_basic.domain.model.TodoItem
+
+interface TodoDataSource {
+    suspend fun getData(id: Int): TodoItem?
+}
