+++ app/src/main/java/com/survivalcoding/network_apps/feature_basic/data/datasource/local/LocalTodoDataSource.kt
@@ -1,0 +1,17 @@
+package com.survivalcoding.network_apps.feature_basic.data.datasource.local
+
+import com.survivalcoding.network_apps.feature_basic.data.datasource.TodoDataSource
+import com.survivalcoding.network_apps.feature_basic.domain.model.TodoItem
+import javax.inject.Singleton
+
+@Singleton
+class LocalTodoDataSource: TodoDataSource {
+    override suspend fun getData(id: Int): TodoItem {
+        return TodoItem(
+            id = 1,
+            title = "test",
+            completed = true,
+            userId = 1
+        )
+    }
+}
