+++ app/src/main/java/com/survivalcoding/network_apps/feature_basic/data/datasource/remote/RemoteTodoDataSource.kt
@@ -1,0 +1,17 @@
+package com.survivalcoding.network_apps.feature_basic.data.datasource.remote
+
+import com.survivalcoding.network_apps.feature_basic.data.datasource.TodoDataSource
+import com.survivalcoding.network_apps.feature_basic.domain.model.TodoItem
+import javax.inject.Inject
+
+class RemoteTodoDataSource @Inject constructor(private val todoService: TodoService) :
+    TodoDataSource {
+    override suspend fun getData(id: Int): TodoItem? {
+        val response = todoService.getTodo(id)
+        return if (response.isSuccessful) {
+            response.body()
+        } else {
+            null
+        }
+    }
+}
