+++ app/src/main/java/com/survivalcoding/network_apps/feature_basic/data/datasource/remote/TodoURL.kt
@@ -1,0 +1,8 @@
+package com.survivalcoding.network_apps.feature_basic.data.datasource.remote
+
+class TodoURL {
+    companion object {
+        const val BASE_URL = "https://jsonplaceholder.typicode.com/"
+        const val TODO_URL = "todos"
+    }
+}
