--- app/src/main/java/com/survivalcoding/network_apps/feature_basic/data/todo/datasource/TodoDataSource.kt
@@ -1,7 +1,0 @@
-package com.survivalcoding.network_apps.feature_basic.data.todo.datasource
-
-import com.survivalcoding.network_apps.feature_basic.domain.model.TodoItem
-
-interface TodoDataSource {
-    suspend fun getData(id: Int): TodoItem?
-}
