--- app/src/main/java/com/survivalcoding/network_apps/feature_basic/data/todo/datasource/remote/TodoURL.kt
@@ -1,8 +1,0 @@
-package com.survivalcoding.network_apps.feature_basic.data.todo.datasource.remote
-
-class TodoURL {
-    companion object {
-        const val BASE_URL = "https://jsonplaceholder.typicode.com/"
-        const val TODO_URL = "todos"
-    }
-}
