--- app/src/main/java/com/survivalcoding/network_apps/feature_basic/domain/repository/ConferenceRepository.kt
@@ -1,7 +1,0 @@
-package com.survivalcoding.network_apps.feature_basic.domain.repository
-
-import com.survivalcoding.network_apps.feature_basic.domain.model.ConferenceItem
-
-interface ConferenceRepository {
-    suspend fun getData(): List<ConferenceItem>?
-}
