--- app/src/main/java/com/survivalcoding/network_apps/feature_basic/presentation/conference/ConferenceActivity.kt
@@ -1,18 +1,0 @@
-package com.survivalcoding.network_apps.feature_basic.presentation.conference
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.survivalcoding.network_apps.databinding.ActivityConferenceBinding
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class ConferenceActivity : AppCompatActivity() {
-    private val binding by lazy {
-        ActivityConferenceBinding.inflate(layoutInflater)
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(binding.root)
-    }
-}
