--- app/src/main/java/com/survivalcoding/network_apps/feature_basic/presentation/conference/adapter/ConferenceListAdapter.kt
@@ -1,20 +1,0 @@
-package com.survivalcoding.network_apps.feature_basic.presentation.conference.adapter
-
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import androidx.recyclerview.widget.ListAdapter
-import com.survivalcoding.network_apps.databinding.ConferenceListItemBinding
-import com.survivalcoding.network_apps.feature_basic.domain.model.ConferenceItem
-
-class ConferenceListAdapter(private val onLinkClicked: (String) -> Any) :
-    ListAdapter<ConferenceItem, ConferenceItemViewHolder>(ConferenceDiffUtilCallback) {
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ConferenceItemViewHolder {
-        val binding =
-            ConferenceListItemBinding.inflate(LayoutInflater.from(parent.context), parent, false)
-        return ConferenceItemViewHolder(binding)
-    }
-
-    override fun onBindViewHolder(holder: ConferenceItemViewHolder, position: Int) {
-        holder.bind(getItem(position), onLinkClicked)
-    }
-}
