+++ app/src/main/res/layout/activity_conference.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature_basic.presentation.conference.ConferenceActivity">
+    tools:context=".conference.presentation.ConferenceActivity">
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragment_container_view"
-        android:name="com.survivalcoding.network_apps.feature_basic.presentation.conference.ConferenceFragment"
+        android:name="com.survivalcoding.network_apps.conference.presentation.ConferenceFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 </androidx.constraintlayout.widget.ConstraintLayout>
