+++ app/src/main/res/layout/basic_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature_basic.presentation.basic.BasicFragment">
+    tools:context=".feature_basic.presentation.BasicFragment">
 
     <TextView
         android:id="@+id/result_text_view"
